--- DotNet/Normalization/Application/Operations/Helpers/CopyOperationHelper.cs
@@ -1,243 +1,0 @@
-﻿using Hl7.Fhir.ElementModel;
-using Hl7.Fhir.FhirPath;
-using Hl7.Fhir.Model;
-using Hl7.FhirPath;
-using System.Collections;
-using System.Reflection;
-
-public class CopyPropertyOperationHelper
-{
-    public static void CopyFhirPathValue(Resource resource, string sourceFhirPath, string targetFhirPath)
-    {
-        if (resource == null || string.IsNullOrEmpty(sourceFhirPath) || string.IsNullOrEmpty(targetFhirPath))
-        {
-            throw new ArgumentException("Resource, SourceFhirPath, and TargetFhirPath must not be null or empty.");
-        }
-
-        // Convert the resource to a navigable element for FHIRPath evaluation
-        var scopedNode = resource.ToTypedElement();
-
-        // Evaluate the source FHIRPath to get the value(s)
-        var sourceValues = scopedNode.Select(sourceFhirPath).ToList();
-
-        if (!sourceValues.Any())
-        {
-            throw new InvalidOperationException($"No values found at source FHIRPath: {sourceFhirPath}");
-        }
-
-        // Take the first value (FHIRPath may return multiple values)
-        var sourceValue = sourceValues.First();
-        var sourcePoco = (sourceValue as ITypedElement)?.ToPoco();
-
-        if (sourcePoco == null)
-        {
-            throw new InvalidOperationException("Source value could not be resolved to a FHIR object.");
-        }
-
-        // Handle primitive and complex types
-        if (sourcePoco is PrimitiveType sourcePrimitive)
-        {
-            // Handle primitive source (e.g., string, integer)
-            if (sourcePrimitive.ObjectValue == null)
-            {
-                throw new InvalidOperationException("Source primitive value is null.");
-            }
-
-            // Set the primitive value at the target path, creating structure if needed
-            SetTargetValue(resource, targetFhirPath, sourcePrimitive.ObjectValue);
-        }
-        else if (sourcePoco is Base sourceComplex)
-        {
-            // Handle complex source by deep copying
-            var copiedObject = sourceComplex.DeepCopy() as Base;
-            SetTargetValue(resource, targetFhirPath, copiedObject);
-        }
-        else
-        {
-            throw new InvalidOperationException("Source type is not supported.");
-        }
-    }
-
-    private static void SetTargetValue(Resource resource, string targetFhirPath, object newValue)
-    {
-        // Split the FHIRPath to identify the parent path and property
-        var pathParts = targetFhirPath.Split('.');
-        if (pathParts.Length < 2)
-        {
-            throw new InvalidOperationException($"Target FHIRPath {targetFhirPath} is too short to resolve parent and property.");
-        }
-
-        // Construct the parent FHIRPath (everything except the last part)
-        var parentPath = string.Join(".", pathParts.Take(pathParts.Length - 1));
-        var propertyName = pathParts.Last();
-
-        // Handle array indexing (e.g., type[0].coding.code)
-        int? arrayIndex = null;
-        if (propertyName.Contains("[") && propertyName.EndsWith("]"))
-        {
-            var indexStart = propertyName.IndexOf('[');
-            var indexEnd = propertyName.IndexOf(']');
-            if (indexStart < indexEnd)
-            {
-                var indexStr = propertyName.Substring(indexStart + 1, indexEnd - indexStart - 1);
-                if (int.TryParse(indexStr, out int index))
-                {
-                    arrayIndex = index;
-                    propertyName = propertyName.Substring(0, indexStart);
-                }
-            }
-        }
-
-        // Evaluate the parent FHIRPath
-        var scopedNode = resource.ToTypedElement();
-        var parentElements = string.IsNullOrEmpty(parentPath) ? [ scopedNode ] : scopedNode.Select(parentPath).ToList();
-
-        Base parentPoco;
-        if (!parentElements.Any())
-        {
-            // Parent path does not exist; create the necessary structure
-            parentPoco = CreateParentStructure(resource, parentPath);
-            if (parentPoco == null)
-            {
-                throw new InvalidOperationException($"Could not create parent structure for {parentPath}.");
-            }
-        }
-        else
-        {
-            // Use the first parent
-            parentPoco = parentElements.First().ToPoco() as Base;
-            if (parentPoco == null)
-            {
-                throw new InvalidOperationException($"Could not resolve parent POCO for {parentPath}.");
-            }
-        }
-
-        // Get the property info
-        var property = parentPoco.GetType().GetProperty(propertyName, BindingFlags.Public | BindingFlags.Instance | BindingFlags.IgnoreCase);
-        if (property == null)
-        {
-            throw new InvalidOperationException($"Property {propertyName} not found on parent type {parentPoco.GetType().Name}.");
-        }
-
-        // Handle single vs. list properties
-        if (typeof(IList).IsAssignableFrom(property.PropertyType))
-        {
-            // Property is a list (e.g., Location.type)
-            var list = property.GetValue(parentPoco) as IList;
-            if (list == null)
-            {
-                list = (IList)Activator.CreateInstance(property.PropertyType);
-                property.SetValue(parentPoco, list);
-            }
-
-            if (arrayIndex.HasValue)
-            {
-                // Ensure the list is large enough
-                while (list.Count <= arrayIndex.Value)
-                {
-                    // Create a new instance of the list item type
-                    var itemType = property.PropertyType.GenericTypeArguments[0];
-                    list.Add(Activator.CreateInstance(itemType));
-                }
-
-                // Set the value at the specified index
-                list[arrayIndex.Value] = newValue;
-            }
-            else
-            {
-                // If no index is specified, append or replace first
-                if (list.Count == 0)
-                {
-                    list.Add(newValue);
-                }
-                else
-                {
-                    list[0] = newValue;
-                }
-            }
-        }
-        else
-        {
-            // Single value property
-            property.SetValue(parentPoco, newValue);
-        }
-    }
-
-    private static Base CreateParentStructure(Resource resource, string parentPath)
-    {
-        if (string.IsNullOrEmpty(parentPath))
-        {
-            return resource;
-        }
-
-        var pathParts = parentPath.Split('.');
-        Base current = resource;
-        foreach (var part in pathParts)
-        {
-            string propertyName = part;
-            int? arrayIndex = null;
-
-            // Handle array indexing (e.g., type[0])
-            if (part.Contains("[") && part.EndsWith("]"))
-            {
-                var indexStart = part.IndexOf('[');
-                var indexEnd = part.IndexOf(']');
-                if (indexStart < indexEnd)
-                {
-                    var indexStr = part.Substring(indexStart + 1, indexEnd - indexStart - 1);
-                    if (int.TryParse(indexStr, out int index))
-                    {
-                        arrayIndex = index;
-                        propertyName = part.Substring(0, indexStart);
-                    }
-                }
-            }
-
-            // Get the property
-            var property = current.GetType().GetProperty(propertyName, BindingFlags.Public | BindingFlags.Instance | BindingFlags.IgnoreCase);
-            if (property == null)
-            {
-                throw new InvalidOperationException($"Property {propertyName} not found on type {current.GetType().Name}.");
-            }
-
-            if (typeof(IList).IsAssignableFrom(property.PropertyType))
-            {
-                // Property is a list
-                var list = property.GetValue(current) as IList;
-                if (list == null)
-                {
-                    list = (IList)Activator.CreateInstance(property.PropertyType);
-                    property.SetValue(current, list);
-                }
-
-                // Ensure the list has enough elements
-                var itemType = property.PropertyType.GenericTypeArguments[0];
-                while (list.Count <= (arrayIndex ?? 0))
-                {
-                    list.Add(Activator.CreateInstance(itemType));
-                }
-
-                // Move to the indexed element
-                current = list[arrayIndex ?? 0] as Base;
-            }
-            else
-            {
-                // Single value property
-                var value = property.GetValue(current) as Base;
-                if (value == null)
-                {
-                    value = Activator.CreateInstance(property.PropertyType) as Base;
-                    property.SetValue(current, value);
-                }
-                current = value;
-            }
-
-            if (current == null)
-            {
-                throw new InvalidOperationException($"Failed to create or access object at path part {part}.");
-            }
-        }
-
-        return current;
-    }
-}
